[BUILD] Add Preserve.cs with [assembly: AlwaysLinkAssembly, Preserve] for a quick, simple way to prevent stripping anything from this package
Indeed, Visual Scripting is no direct code reference and custom units used in script graphs, present in external Packages would get stripped even at stripping level Low.

Usually we'd add this assembly to link.xml in each project using it, but since there is only one custom unit now, and even if we add some, there won't be many generic ones to add to this package, we assume that all projects importing this package want to use most of the custom units there, so not stripping them is fine.