[CURVES 2D] Bezier: improved split point: it now moves continuously along the Bezier path, and we exclude edge positions
We also make the key point handles bigger when hovering them, to show we are moving them, not adding a split point nearby.

To get a smooth split point, we projected the cursor position to the closest segment to get the nearest point, computed the corresponding curve parameter ratio and called SplitCurveAtParameterRatio on click.