[AUDIO] Extracted BaseSfxPoolManager from SfxPoolManager, renamed latter InGameSfxPoolManager, added UISfxPoolManager
This allows us to have multiple Sfx Pool Managers, one per sound category (this was impossible before as SfxPoolManager was a concrete singleton). Each project can even define more categories and pool managers by subclassing BaseSfxPoolManager and using the CRTP.

This should be retrocompatible with existing projects' manager prefabs thanks to renaming SfxPoolManager -> InGameSfxPoolManager preserving .meta file. However, code itself won't be compatible and all occurrences of SfxPoolManager should be renamed InGameSfxPoolManager (or UISfxPoolManager if that project wants to use it).

Since the scripts are now different, we cannot just have a SfxPoolManager_Base.prefab so this was deprecated with 'OLD_' prefix, again to preserve retrocompatibility for projects using a variant of it. But those projects should eventually replace the prefab variant with brand new, unrelated prefabs that have either InGameSfxPoolManager or UISfxPoolManager component.