[SHADER] Added experimental UI-Default-with-Brightness.shader -> MAT_UI_Default-with-Brightness.mat inspired by Sprite-Unlit.shadergraph to allow brightening UI graphics
! We found graphical glitches when testing various UI sprites, even when checking for pure white. It also doesn't work with a Mask !

but in order to work with UI (Images, etc.), this needed Stencil properties which were too hard to make work via shader graph by hand. So instead, we copied builtin_shaders-2021.1.15f1/DefaultResourcesExtra/UI/UI-Default.shader and added a _Brightness property.

We wrote two implementations: A. (commented out) which is the written version of Sprite-Unlit.shadergraph, and B. (active) which uses HSL instead of HSV and gives even better results. Consider reinjecting this version in Sprite-Until shadergraph, or even Sprite-Unlit written shader if it makes it easier to copy.

This means that ImageMaterialPropertyController.cs can now work with MAT_UI_Default-with-Brightness.mat but results will be slightly different than SpriteMaterialPropertyController.cs with MAT_Sprite-Unlit.