[DEBUG] Added experimental DetectRectTransformOverride + Flag component, but it has too many false positive (although it does find the issues in the pack)
This is meant to identify the infamous "shrinked UI prefab instances", esp. the pooled ones pre-placed in scenes so only the first instances of those will have bad dimensions, sometimes shrinked to 0 width/height!

But for now, project developers should not rely on this and instead place their pooled UI objects (and other objects prone to bad prefab overrides being reused) in some EditorOnly _WorkshopCanvas so we are sure that at runtime, brand new objects are instantiated from the prefabs, with no overrides.