[SETTINGS] (! DATA BREAKING for VolumeSettingData !) New engine/custom setting system
We now only store custom settings (ICustomSetting) and directly get/set values in engine state for IEngineSetting.

Replaced GetDefaultValueOnStart with ICustomSetting.GetDefaultValue and IEngineSetting.GetValue. When dealing with engine setting, do not get/set value in dictionary anymore, and just use IEngineSetting.Get/SetValue directly.

Replaced OnSetValue with IEngineSetting.GetValue (custom settings has no side effects on set).

Removed alwaysCheckEngineValue flag since IEngineSetting is enough to distinguish engine settings, and all work directly with engine values now.

Note that we lost the notion of "default" for engine settings (we could store the initial value at least, but even then it's not a true default if player has started changing values and the engine remembers next session). We may need to restore a default value for engines later, but for now it's enough to work when there are no preferences.

ContinuousSettingData is now aware of its range, so VolumeSettingData uses it, but sets it to [-80, 0] on Reset. This won't affect existing data, so existing projects must set range min to -80 manually.