[Build] Extract const string BuildProfilesFolder and public static string[] GetBuildProfileGUIDs()
In the end, we didn't use it in other scripts: it was meant to create a UITK Editor Window that would fill a dropdown with the result of GetBuildProfileGUIDs, but it turned out that I had to store and display Build Profile full paths, and then remap them to the assets, which was a bit cumbersome. In the end I decided to use a classic EditorGUILayout.ObjectField which was much faster.

But I kept the refactoring which is not bad either.