[Editor Run Command] Add EditorRunCommand.RunCommand static method to run an arbitrary bash script
It works on Windows too as long as bash is in PATH.

Limitation: Due to bash executing a script, we cannot say if the script is missing or not easily (we could File.Exists with full path but it's more complex to determine if file exists in any PATH). We can only notice that we debug log an empty message (but if script was silent, we cannot distinguish a success from a failure).